Update dependencies, add Utiliy resource enhancer, use providers
