+++ radar-jersey/src/main/kotlin/org/radarbase/jersey/config/GeneralExceptionResourceEnhancer.kt
@@ -6,6 +6,7 @@ import org.radarbase.jersey.exception.mapper.WebApplicationExceptionMapper
 /** Add WebApplicationException and any exception handling. */
 class GeneralExceptionResourceEnhancer: JerseyResourceEnhancer {
     override val classes: Array<Class<*>> = arrayOf(
-            UnhandledExceptionMapper::class.java,
-            WebApplicationExceptionMapper::class.java)
+        UnhandledExceptionMapper::class.java,
+        WebApplicationExceptionMapper::class.java,
+    )
 }
