+++ radar-jersey/src/main/kotlin/org/radarbase/jersey/doc/swagger/SwaggerResourceEnhancer.kt
@@ -6,7 +6,6 @@ import io.swagger.v3.oas.models.OpenAPI
 import org.glassfish.jersey.server.ResourceConfig
 import org.radarbase.jersey.config.JerseyResourceEnhancer
 
-
 class SwaggerResourceEnhancer(
     private val openApi: OpenAPI,
     private val ignoredRoutes: Set<String>? = null,
