+++ radar-jersey/src/main/resources/logback.xml
@@ -31,8 +31,6 @@
         <appender-ref ref="AUTH" />
     </logger>
 
-    <logger name="org.glassfish.jersey.internal.inject.Providers" level="SEVERE"/>
-
     <root level="INFO">
         <appender-ref ref="STDOUT" />
     </root>
